.evergreen/config_generator/requirements.txt
pydantic~=1.10
PyYAML~=5.4
shrub.py~=3.0
yamlordereddictloader~=0.4
